refactor: extract allocation auth helpers to dedicated module
Move `authenticate_api_request` and `requires_allocation_auth` from
`views/__init__.py` (already 1000+ lines) into a dedicated
`views/allocation_auth.py`. Pure move with no behaviour change —
prep work for upcoming signature-based auth, isolated here so the
crypto change lands as a focused diff.

- New: `views/allocation_auth.py` hosts both helpers.
- `views/__init__.py` re-imports `authenticate_api_request` for the
  three direct call sites (`update_allocations`, `recreate_network`,
  `regenerate_proxy`); the unused `sha256` and the dropped
  decorator's `functools` are tidied accordingly.
- `views/migration.py` imports the decorator from the new module.
- `tests/supervisor/views/test_migration.py` updates its 4 patch
  targets (and the explanatory docstring) to point at
  `allocation_auth.authenticate_api_request`, since the decorator
  now resolves the name in its own module's namespace.
